The dict-bindings in an IPBinds need not be in dependency order
This appears to be a long-standing bug, discovered by BlueSpec 
(ravi@bluespec.com), Trac bug #795

The problem was that in an IP binding group, the dict bindings
aren't necessarily in dependency order; and if they aren't 
we get a core-lint error.

Test tc203 checks this case.  (Though whether it shows up at
all depends a bit on accidental factors of binding ordering.)