Disable form-checking for rule LHSs
Previously we checked the form of the arguments of a RULE lhs, to 
ensure that they were simple applications. There was no good reason
for that, save to prevent you writing LHSs that were unlikely to match.

And Don Stewart found he wanted to do something we didn't allow (a section,
I think).  So I have just disabled the check.