Result type signatures are no longer supported (partial)
I had failed to remove the bit where result type signatures bind
lexical type variables.  And now we are planning to remove them entirely.

This commit therefore does a partial removal (to avoid destablising 6.6).
It also arranges that
	f :: sig = rhs
means a *pattern* binding (not a function binding with no arguments 
and a result signature), which makes sense.