Fix a long-standing but obscure bug in worker-wrapper generation
Worker/wrapper generation sometimes has to add a dummy void (State#) argument
to retain laziness.  But when generating the strictness signature for the
worker, I forgot to take account of the extra argument, resulting in a
bogus strictness signature.

Result, chaos.  Trac 317 shows this up, and this patch fixes it.