Make SpecConstr work right for nullary constructors
For totally stupid reasons, SpecConstr didn't work for the (particularly
easy) case of nullary constructors like True and False.  I just had some
equations in the wrong order, so that a Var case came first, which 
matches a nullary constructor, with the constructor-application case
afterwards.

The fix is easy.  I did a bit of refactoring at the same time.