MERGE: In hashExpr, use Word32 rather than relying on wrapping behaviour of Int Fixes #952, as it turns out.
When compiling via C, we are at the mercy of C's undefined behaviour
with respect to overflow of signed integer operations, and this was
biting us here.

Perhaps we should always add the -fwrapv flag to gcc, but since
Haskell doesn't define overflow on Int either, it seemed the right
thing to do to fix this code anyway.